Refactored a bit.
