+++ server/custom_kernels/setup.py
@@ -1,5 +1,5 @@
 from setuptools import setup
-from torch.utils.cpp_extension import BuildExtension, CUDAExtension, CppExtension
+from torch.utils.cpp_extension import BuildExtension, CUDAExtension
 
 setup(
     name="custom_kernels",
@@ -14,7 +14,7 @@ setup(
             sources=["custom_kernels/fused_attention_cuda.cu"],
             extra_compile_args=["-arch=compute_80", "-std=c++17"],
         ),
-        CppExtension(
+        CUDAExtension(
             name="custom_kernels.exllama",
             sources=[
                 "custom_kernels/exllama/exllama_ext.cpp",
