+++ server/text_generation_server/models/flash_santacoder.py
@@ -66,7 +66,6 @@ class FlashSantacoderSharded(FlashCausalLM):
         super(FlashSantacoderSharded, self).__init__(
             model=model.to(device),
             tokenizer=tokenizer,
-            config=config,
             num_layers=len(model.transformer.h),
             num_kv_heads=1,
             head_size=model.transformer.head_size,
