+++ server/text_generation_server/models/gpt_neox.py
@@ -63,7 +63,6 @@ class GPTNeoxSharded(CausalLM):
         super(CausalLM, self).__init__(
             model=model,
             tokenizer=tokenizer,
-            config=config,
             requires_padding=True,
             dtype=dtype,
             device=device,
