+++ server/text_generation_server/models/model.py
@@ -8,14 +8,6 @@ from transformers import PreTrainedTokenizerBase, PretrainedConfig
 from text_generation_server.models.types import Batch, GeneratedText
 from text_generation_server.pb.generate_pb2 import InfoResponse
 
-from text_generation_server.utils.gptq.quant_linear import Ex4bitLinear
-from custom_kernels.exllama import prepare_buffers, set_tuning_params
-
-from text_generation_server.utils.layers import (
-    TensorParallelRowLinear,
-    TensorParallelColumnLinear
-)
-
 B = TypeVar("B", bound=Batch)
 
 class Model(ABC):
@@ -23,7 +15,6 @@ class Model(ABC):
         self,
         model: torch.nn.Module,
         tokenizer: PreTrainedTokenizerBase,
-        config: PretrainedConfig,
         requires_padding: bool,
         dtype: torch.dtype,
         device: torch.device,
@@ -46,47 +37,6 @@ class Model(ABC):
             inspect.signature(model.forward).parameters.get("position_ids", None)
             is not None
         )
-        self.config = config
-
-        if config.quantize == "gptq":
-            # Buffers need to be persistent to avoid any bug.
-            self.buffers = {}
-            use_exllama_act_order = False
-            max_dq_buffer_size = 1
-            max_inner_outer_dim = 1
-            for name, submodule in model.named_modules():
-                if isinstance(submodule, (TensorParallelColumnLinear, TensorParallelRowLinear)) and isinstance(submodule.linear, Ex4bitLinear):
-
-                    max_dq_buffer_size = max(max_dq_buffer_size, submodule.linear.qweight.numel() * 8)
-
-                    if submodule.linear.act_order:
-                        max_inner_outer_dim = max(max_inner_outer_dim, submodule.linear.height, submodule.linear.width)
-
-                        use_exllama_act_order = True
-                                    
-            if use_exllama_act_order:
-                # TODO: this should be set to rust side `max_total_tokens`, but TGI
-                # does not offer an API to expose this variable to python, as this variable
-                # is handled by the client but it appears the model is initialized by the server.
-                # An alternative could be to initialize the buffers during warmup.
-                max_total_tokens = 2048
-            else:
-                max_total_tokens = 1
-
-            # This temp_state buffer is required to reorder X in the act-order case.
-            self.buffers["temp_state"] = torch.zeros((max_total_tokens, max_inner_outer_dim), dtype=torch.float16, device=device)
-
-            # This temp_dq buffer is required to dequantize weights when using cuBLAS, typically for the prefill.
-            self.buffers["temp_dq"] = torch.zeros((1, max_dq_buffer_size), dtype=torch.float16, device=device)
-            
-            prepare_buffers(device, self.buffers["temp_state"], self.buffers["temp_dq"])
-
-            matmul_recons_thd = 8
-            matmul_fused_remap = False
-            matmul_no_half2 = False
-            set_tuning_params(matmul_recons_thd, matmul_fused_remap, matmul_no_half2)
-
-            torch.cuda.empty_cache()
 
         self.check_initialized()
