+++ server/text_generation_server/models/opt.py
@@ -61,7 +61,6 @@ class OPTSharded(CausalLM):
         super(CausalLM, self).__init__(
             model=model,
             tokenizer=tokenizer,
-            config=config,
             requires_padding=True,
             dtype=dtype,
             device=device,
