+++ server/text_generation_server/models/rw.py
@@ -58,7 +58,6 @@ class RW(CausalLM):
         super(CausalLM, self).__init__(
             model=model,
             tokenizer=tokenizer,
-            config=model.config,
             requires_padding=True,
             dtype=dtype,
             device=device,
