+++ server/text_generation_server/models/santacoder.py
@@ -63,7 +63,6 @@ class SantaCoder(CausalLM):
         super(CausalLM, self).__init__(
             model=model,
             tokenizer=tokenizer,
-            config=model.config,
             requires_padding=True,
             dtype=dtype,
             device=device,
