+++ server/text_generation_server/models/seq2seq_lm.py
@@ -542,7 +542,6 @@ class Seq2SeqLM(Model):
         super(Seq2SeqLM, self).__init__(
             model=model,
             tokenizer=tokenizer,
-            config=model.config,
             requires_padding=True,
             dtype=dtype,
             device=device,
